# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Backend Base NestJS
  version: '1.0'
  description: Documentación Backend base - NestJS
  contact:
    name: agetic@agetic.gob.bo
servers:
  - url: 'https://base.agcs.agetic.gob.bo/ws/api'
paths:
...
  /autorizacion/roles:
    get:
      operationId: RolController_listar
      summary: API para obtener el listado de Roles
      parameters: []
      responses:
        '200':
          description: ''
      tags: &ref_2
        - Roles
      security: &ref_3
        - bearer: []
    post:
      operationId: RolController_crear
      summary: API para crear un rol
      parameters: []
      requestBody:
        required: true
        description: new Rol
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CrearRolDto'
      responses:
        '201':
          description: ''
      tags: *ref_2
      security: *ref_3
  /autorizacion/roles/todos:
    get:
      operationId: RolController_listarTodos
      summary: API para obtener el listado de todos los Roles
      parameters:
        - name: limite
          required: false
          in: query
          schema:
            minimum: 10
            maximum: 50
            default: 10
            type: number
        - name: pagina
          required: false
          in: query
          schema:
            minimum: 1
            default: 1
            type: number
        - name: filtro
          required: false
          in: query
          schema:
            type: string
        - name: ordenRaw
          required: false
          in: query
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_2
      security: *ref_3
  /autorizacion/roles/{id}:
    patch:
      operationId: RolController_actualizar
      summary: API para actualizar un rol
      parameters:
        - name: id
          required: true
          in: path
          example: '12'
          schema:
            type: string
      requestBody:
        required: true
        description: new Rol
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActualizarRolDto'
      responses:
        '200':
          description: ''
      tags: *ref_2
      security: *ref_3
  /autorizacion/roles/{id}/activacion:
    patch:
      operationId: RolController_activar
      summary: API para activar un rol
      parameters:
        - name: id
          required: true
          in: path
          example: '12'
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_2
      security: *ref_3
  /autorizacion/roles/{id}/inactivacion:
    patch:
      operationId: RolController_inactivar
      summary: API para inactivar un rol
      parameters:
        - name: id
          required: true
          in: path
          example: '12'
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_2
      security: *ref_3
  /autorizacion/modulos:
    get:
      operationId: ModuloController_listar
      summary: API para obtener el listado de Módulos
      parameters:
        - name: limite
          required: false
          in: query
          schema:
            minimum: 10
            maximum: 50
            default: 10
            type: number
        - name: pagina
          required: false
          in: query
          schema:
            minimum: 1
            default: 1
            type: number
        - name: filtro
          required: false
          in: query
          schema:
            type: string
        - name: ordenRaw
          required: false
          in: query
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: &ref_4
        - Módulos
      security: &ref_5
        - bearer: []
    post:
      operationId: ModuloController_crear
      summary: API para crear un Módulo
      parameters: []
      requestBody:
        required: true
        description: new Modulo
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CrearModuloDto'
      responses:
        '201':
          description: ''
      tags: *ref_4
      security: *ref_5
    delete:
      operationId: ModuloController_eliminar
      summary: API para eliminar un Módulo
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_4
      security: *ref_5
  /autorizacion/modulos/{id}:
    patch:
      operationId: ModuloController_actualizar
      summary: API para actualizar un Módulo
      parameters:
        - name: id
          required: true
          in: path
          example: '12'
          schema:
            type: string
      requestBody:
        required: true
        description: Modulo
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActualizarModuloDto'
      responses:
        '200':
          description: ''
      tags: *ref_4
      security: *ref_5
  /autorizacion/modulos/{id}/activacion:
    patch:
      operationId: ModuloController_activar
      summary: API para activar un Módulo
      parameters:
        - name: id
          required: true
          in: path
          example: '12'
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_4
      security: *ref_5
  /autorizacion/modulos/{id}/inactivacion:
    patch:
      operationId: ModuloController_inactivar
      summary: API para inactivar un Módulo
      parameters:
        - name: id
          required: true
          in: path
          example: '12'
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_4
      security: *ref_5
...
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    CrearRolDto:
      type: object
      properties:
        rol:
          type: string
          example: CONSULTA
        nombre:
          type: string
          example: Consulta
        estado:
          type: string
          example: ACTIVO
      required:
        - rol
        - nombre
        - estado
    ActualizarRolDto:
      type: object
      properties:
        rol:
          type: string
          example: ADMINISTRADOR
        nombre:
          type: string
          example: Administrador
        estado:
          type: string
          example: ACTIVO
      required:
        - rol
        - nombre
        - estado
    PropiedadesDto:
      type: object
      properties:
        icono:
          type: string
          example: dashboard
        descripcion:
          type: string
          example: Módulo de estadísticas
        orden:
          type: number
          example: 7
      required:
        - icono
        - descripcion
        - orden
    CrearModuloDto:
      type: object
      properties:
        label:
          type: string
          example: Estadísticas
        url:
          type: string
          example: /admin/estadisticas
        nombre:
          type: string
          example: Estadísticas
        propiedades:
          $ref: '#/components/schemas/PropiedadesDto'
        estado:
          type: string
          example: ACTIVO
      required:
        - label
        - url
        - nombre
        - propiedades
        - estado
    ActualizarModuloDto:
      type: object
      properties:
        label:
          type: string
          example: Trámites
        url:
          type: string
          example: /admin/tramites
        nombre:
          type: string
          example: Módulo de trámites
        propiedades:
          $ref: '#/components/schemas/PropiedadesDto'
        estado:
          type: string
          example: ACTIVO
      required:
        - label
        - url
        - nombre
        - propiedades
        - estado
...